Shared: move custom-scrollbar from utils to scrollbar component
