+++ packages/shared/components/scrollbar/custom-scrollbar/ScrollbarThumb.tsx
