+++ packages/shared/components/scrollbar/custom-scrollbar/style.ts
